add some error handling

+++ scripts/timeout-processing.lua
@@ -11,28 +11,36 @@ local processing = redis.call('hgetall', processingHash)
 
 for i=1, #processing, 2 do
 
-	local taskInfo = cjson.decode(processing[i+1])
+	local taskInfo = nil;
 
-	if (taskInfo['metadata']['started'] + taskInfo['metadata']['timeout']) < timestamp then
+	pcall(function()
+		taskInfo = cjson.decode(processing[i+1])
+	end)
+	
+	if type(taskInfo) == 'table' and taskInfo['metadata'] then
 
-		taskInfo['metadata']['attempts'] = taskInfo['metadata']['attempts'] - 1 
-		local task = taskInfo['metadata']['task']
+		if (taskInfo['metadata']['started'] + taskInfo['metadata']['timeout']) < timestamp then
 
-		if taskInfo['metadata']['attempts'] > 0 then
-			--add to released list again
-			taskInfo['metadata']['started'] = nil
-			redis.call('rpush', releasedList, cjson.encode(taskInfo))
+			taskInfo['metadata']['attempts'] = taskInfo['metadata']['attempts'] - 1 
+			local task = taskInfo['metadata']['task']
 
-			redis.call('hset', closedHash, task, cjson.encode(taskInfo))
-		else
-			--add to failed list
-			redis.call('hdel', closedHash, task)
+			if taskInfo['metadata']['attempts'] > 0 then
+				--add to released list again
+				taskInfo['metadata']['started'] = nil
+				redis.call('rpush', releasedList, cjson.encode(taskInfo))
 
-			local taskInfo = {task=task, failed=timestamp}
-			redis.call('rpush', failedList, cjson.encode(taskInfo))
+				redis.call('hset', closedHash, task, cjson.encode(taskInfo))
+			else
+				--add to failed list
+				redis.call('hdel', closedHash, task)
+
+				local taskInfo = {task=task, failed=timestamp}
+				redis.call('rpush', failedList, cjson.encode(taskInfo))
+			end
+
+			--delete from processing
+			redis.call('hdel', processingHash, task)
 		end
-
-		--delete from processing
-		redis.call('hdel', processingHash, task)
 	end
+
 end
